Use scope borrows
