+++ program/src/processor/initialize_mint2.rs
@@ -1,11 +1,11 @@
 use pinocchio::{account_info::AccountInfo, ProgramResult};
 
-use super::initialize_mint::process_initialize_mint;
+use super::shared;
 
 #[inline(always)]
 pub fn process_initialize_mint2(
     accounts: &[AccountInfo],
     instruction_data: &[u8],
 ) -> ProgramResult {
-    process_initialize_mint(accounts, instruction_data, false)
+    shared::initialize_mint::process_initialize_mint(accounts, instruction_data, false)
 }
