+++ program/src/processor/shared/approve.rs
@@ -44,30 +44,38 @@ pub fn process_approve(
         };
 
     // Validates source account.
+    {
+        // SAFETY: scoped immutable borrow of `source_account_info` account data. When
+        // `owner_info` is the same as `source_account_info`, there will be another immutable
+        // borrow in `validate_owner` – this is safe because both borrows are immutable.
+        let source_account =
+            unsafe { load::<Account>(source_account_info.borrow_data_unchecked())? };
 
-    let source_account =
-        unsafe { load_mut::<Account>(source_account_info.borrow_mut_data_unchecked())? };
-
-    if source_account.is_frozen() {
-        return Err(TokenError::AccountFrozen.into());
-    }
-
-    if let Some((mint_info, expected_decimals)) = expected_mint_info {
-        if mint_info.key() != &source_account.mint {
-            return Err(TokenError::MintMismatch.into());
+        if source_account.is_frozen() {
+            return Err(TokenError::AccountFrozen.into());
         }
 
-        let mint = unsafe { load::<Mint>(mint_info.borrow_data_unchecked())? };
+        if let Some((mint_info, expected_decimals)) = expected_mint_info {
+            if mint_info.key() != &source_account.mint {
+                return Err(TokenError::MintMismatch.into());
+            }
 
-        if expected_decimals != mint.decimals {
-            return Err(TokenError::MintDecimalsMismatch.into());
+            let mint = unsafe { load::<Mint>(mint_info.borrow_data_unchecked())? };
+
+            if expected_decimals != mint.decimals {
+                return Err(TokenError::MintDecimalsMismatch.into());
+            }
         }
-    }
 
-    validate_owner(&source_account.owner, owner_info, remaining)?;
+        validate_owner(&source_account.owner, owner_info, remaining)?;
+    }
 
     // Sets the delegate and delegated amount.
 
+    // SAFETY: any immutable borrow of `source_account_info` account data is dropped at
+    // this point, so it is safe to borrow mutably.
+    let source_account =
+        unsafe { load_mut::<Account>(source_account_info.borrow_mut_data_unchecked())? };
     source_account.set_delegate(delegate_info.key());
     source_account.set_delegated_amount(amount);
