+++ program/src/processor/shared/mod.rs
@@ -6,6 +6,7 @@
 pub mod approve;
 pub mod burn;
 pub mod initialize_account;
+pub mod initialize_mint;
 pub mod initialize_multisig;
 pub mod mint_to;
 pub mod toggle_account_state;
